requirements.txt
black==24.10.0
pymongo
python-dotenv